# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "dailyReadings API",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "localhost:50000",
  "basePath": "/",
  "tags": [
    {
      "name": "dailyReadings",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "API_KEY",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/dailyReading": {
      "post": {
        "tags": [
          "dailyReadings"
        ],
        "description": "Endpoint to store a new dailyReading.",
        "parameters": [
          {
            "name": "requestBody",
            "in": "body",
            "description": "Input data.",
            "schema": {
              "$ref": "#/definitions/dailyReadingInput"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/dailyReading"
            },
            "description": "Response after successful post."
          }
        }
      },
      "put": {
        "tags": [
          "dailyReadings"
        ],
        "description": "Endpoint to update the dailyReading.",
        "parameters": [
          {
            "name": "readingID",
            "in": "query",
            "type": "string"
          },
          {
            "name": "requestBody",
            "in": "body",
            "description": "Input data.",
            "schema": {
              "$ref": "#/definitions/dailyReadingInput"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/dailyReading"
            },
            "description": "Response after successful update."
          }
        }
      },
      "delete": {
        "tags": [
          "dailyReadings"
        ],
        "description": "Endpoint to delete one dailyReading.",
        "parameters": [
          {
            "name": "readingID",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/dailyReadingWasDeleted"
            },
            "description": "Response after successful delete."
          }
        }
      }
    },
    "/dailyReadings": {
      "get": {
        "tags": [
          "dailyReadings"
        ],
        "description": "Endpoint to get the dailyReadings.",
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/dailyReadingsOutput"
            },
            "description": "Returning data."
          }
        }
      },
      "post": {
        "tags": [
          "dailyReadings"
        ],
        "description": "Endpoint to store a couple of new dailyReadings.",
        "parameters": [
          {
            "name": "requestBody",
            "in": "body",
            "description": "Input data.",
            "schema": {
              "$ref": "#/definitions/dailyReadingsInput"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/dailyReadingsOutput"
            },
            "description": "Returning data."
          }
        }
      }
    },
    "/dailyReading/{year}/{month}/{day}": {
      "get": {
        "tags": [
          "dailyReadings"
        ],
        "description": "Endpoint to get the dailyReading.",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "day",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/dailyReading"
            },
            "description": "Returning data."
          }
        }
      }
    }
  },
  "definitions": {
    "dailyReading": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "606abafb977eff055b52505a"
        },
        "dailyReading": {
          "type": "string",
          "example": "test reading"
        },
        "readingYear": {
          "type": "string",
          "example": "2021"
        },
        "readingMonth": {
          "type": "string",
          "example": "04"
        },
        "readingDay": {
          "type": "string",
          "example": "01"
        },
        "created_by": {
          "type": "string",
          "example": "Admin"
        },
        "created_date": {
          "type": "string",
          "example": "021-04-05T07:23:39.313Z"
        },
        "status": {
          "type": "string",
          "example": "saved"
        }
      }
    },
    "dailyReadingsInput": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/dailyReadingInput"
      }
    },
    "dailyReadingsOutput": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/dailyReading"
      }
    },
    "dailyReadingInput": {
      "type": "object",
      "properties": {
        "dailyReading": {
          "type": "string",
          "example": "test reading"
        },
        "readingYear": {
          "type": "string",
          "example": "2021"
        },
        "readingMonth": {
          "type": "string",
          "example": "04"
        },
        "readingDay": {
          "type": "string",
          "example": "01"
        },
        "created_by": {
          "type": "string",
          "example": "Admin"
        }
      }
    },
    "dailyReadingWasDeleted": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Reading successfully deleted"
        }
      }
    }
  }
}